backend/requirements.txt
fastapi
uvicorn
telethon
python-dotenv
passlib[bcrypt]
fastapi-jwt-auth
